fix: normalizedRequestAttrs takes a dasherized model name
closes #1440
